Authorization

+++ app/Http/Controllers/PostController.php
@@ -9,7 +9,7 @@ class PostController extends Controller
 {
     public function index()
     {
-        $posts = Post::with(['user', 'likes'])->paginate(20);
+        $posts = Post::latest()->with(['user', 'likes'])->paginate(20);
 
         return view('posts.index', [
             'posts' => $posts
@@ -29,13 +29,10 @@ class PostController extends Controller
 
     public function destroy(Post $post)
     {
+
+        $this->authorize('delete', $post);
         $post->delete();
         return back();
     }
-    //public function destroy(Post $post)
-    //{
-    //    $post->delete();
-    //    return redirect()->route('$post.index')
-    //        ->with('success', 'Post was deleted successfully');
-    //}
+
 }
